Catapult AI using Assembly
This project uses a car with a claw that can throw spherical projectiles into cups. It is up to the user to use the keyboard to move the vehicle, and the vehicle will decide and attepmt to score on the net that's most likely to be scored on. There is a 60 second time limit; and if you score 5 times, you win.